created hook folder and moved searhAPI over
